Eliminate SER_KEEP flag (reclaim for SER_FRAME)
Given the packed nature of series there is only room for 8 bit flags,
and they were all used up...so it was time to get around to culling
SER_KEEP in order to add SER_FRAME.

SER_KEEP was a flag available throughthe RL_Api to ask for an
indefinite lock on a series, undermining the garbage collector.  It
happened to also be used for setting "keep" status on elements in
the root set, which is unnecessary because they are in the root set.

This deprecated the RL_Api (which should use a GC hook regarding
a list it keeps and still provably has values to, instead of a bit that
can "leak").  It also did the necessary changes to permit the root
set elements to go ahead and have "managed" status and be put
under GC.