Add ANY-SERIES/SCALAR/NUMBER to r2r3-future.reb
The r2r3-future.reb file is something of relatively limited interest (for
adapting Rebol2 and R3-Alpha in a way that it might run Ren/C or future
scripts).  That will be impossible in general, but for a limited sense it can
work for a subset of files.  One relevant subset: the make prep scripts.

Because FAIL is a widely relevant change, it has a small shim in the
r2r3-future.r file.  The FAIL shim depended on scalar? which was converted
to any-scalar?, which had no shim.  This adds those shims, and makes a
note that this stub file needs a "Czar" to get it into better shape for those
who would be interested beyond the `make prep` script's needs.