ANY-TYPE! => ANY-VALUE!
There is an ambiguity inherent when types themselves are values
in the language, and you say "ANY-TYPE".  The problem is that may
suggest something more narrow than what has been meant, which
historically is that ANY-TYPE meant "any legal Rebol value" (and that
included UNSET!, when UNSET! was a legal Rebol value, which it is
now planned to not be).

Renaming it to ANY-VALUE! clears up the problem, so you don't think
it suggests it could be "any type", e.g. STRING! or INTEGER! etc.

ANY-TYPE! added to legacy for backward compatibilty.