Cleanups and enhancements of Ren/C zlib usage
In the early commits of Ren/C it absorbed a PR that had been pending
on rebol/rebol's GitHub which brought Rebol up to date with a copy of
zlib that was extracted using a script.  (The old copyright date was 1998,
and there had been a number of fixes between then and 2013.)

This updated library had not been tested with the unzip.reb script, which
was failing to work.  Reasoning was that zlib had been "hacked" to
accept an invalid format.  Updating had removed that hack:

    http://stackoverflow.com/q/32907420/211160

In order to support correct function of the library, this adds an /ONLY
refinement to COMPRESS and DECOMPRESS.  It also in general solves
a number of problems, e.g. making GZIP reading and writing functional
so that you can process .GZ files.  Other adjustments included based on
following modern Zlib documentation.

This change requires an updated version of unzip.reb to work, as the old
version was doing something rather incorrect.