ERROR => same struct as OBJECT/PORT/MODULE
At one point in time, ERROR! objects did double duty as the value
type for THROWN() values generated for function returns and
similar.  The error then contained a number identifying the throw
and a possible argument.

With the desire to allow a "full value's worth" of value to /NAME a
throw, the THROWN() convention was changed to a header bit that
could be set on any value type.  This meant all errors were changed
to have "object frames" and be ordinary objects which could be
handed to the user (with two exceptions: halt and a parse hack).

This commit started as comment maintenance but then shifted to
merging ERROR! in to be in the same form as PORT!, MODULE!,
and OBJECT! which share a common REBVAL structure.  (This
is similar to how series values like BLOCK! and STRING! and
PAREN! and BINARY! share the same structure.)