Adjustments for unknown-size decompression experiment
The primary use-case for the change to the zlib interface was
unzip.reb, which knew the size of the raw decompression...but
some not-very-tested code was added to kick off an attempt at
being able to inflate a block of deflate data without storing or
stating a size.  It had an error in that it expanded the target
series, but if it had to reallocate the series then the underlying
pointer would not adjust, so zlib had to be told where the new
pointer could be.

Fixes a crash on an error-reporting condition, as well as a
potential memory leak from zlib internal structures on error.