WORD! => SET-WORD! instances in %errors.r
The bootstrap load of data doesn't do checking on the object it loads
(no binding of words) so it did not notice that two WORD! instances
had crept into where set-words should be for error keys.

This fixes that and does a cleanup and more checking in make-boot.r
to prevent such things from happening again.  It puts the make-boot
into the same terminology as the error code itself to make it easier
to understand and modify.  It also uses FAIL which has been added
to r2r3-future.r.

Because it was tempting to use EVERY instead of FOR-EACH, this
*removes* every from the past-to-future Rebol adaptation, because
there is no way to write a working loop wrapper in R3-Alpha.  Hence
in make-boot modifications it should not be used, even if it looks
nicer...and making it a synonym is now a lie as EVERY has a different
(and more useful) return result.