Ensure error argument NULL termination size
A check on error creation ensured that all errors originating from
C code were created with the right number of arguments for their
template message.  This adjusts cases where NULL was passed
in and shouldn't have been, too many nulls were given, or moves
after-the-fact argument supply to be before-the-fact to avoid
triggering the alarm.