!Make_Error_Object(...) => Make_Error_Object_Throws(...)
At one point Make_Error_Object was retrofitted to "bubble up" the
cases where a throw happened in the evaluation, and this was
passed as a "success" or "failure" flag.  Overall in the system as more
such leaks were plugged the pattern emerged that usually the caller
needed to test the throw condition and do handling...and it was not
that clear what a boolean result meant at the callsite.  By inverting the
logic and adding "throws" into the name it could be read more easily
as testing whether the operation in the if statement is (in the present
tense) throwing during the conditional test.

This propagates that rule to Make_Error_Object, and lessens the need
for the "THROWN()" assertions at the callsites to document the meaning
of the branch.