Roll back speculative change of frame offset in cell
When all OBJECT! type structures were unified under Reb_Object
in c48b9d5, it led to the removal of an assertion that Reb_Error and
Reb_Object had their frames at the same offset in a value cell.
(as the same structure, it was now impossible for them to be different).

This suggested checking to see if the common position of the frame
at the same pointer location was essential or not.  If it was not, then
it would be possible to make changes to the behavior of ANY-OBJECT!
or ANY-SERIES! with one not affecting the other by having the same
requirements of behavior w.r.t. VAL_SERIES and VAL_OBJ_FRAME.

At least one issue was found with an assumption in core code itself
(patched here--caused defects in deep copying apparently not checked
by any tests).  Yet the reason for undoing the change anyway is due
to the value cell layouts being mirrored and dependent in the RXIARG
structure.  A simple patch of the address in the macros did not fix it,
suggesting that series RXIARGs and object RXIARGs are being called
randomly interchangeable.

Rather than investigate the specific cause of that bug, the speculative
change is rolled back here and an assertion and notes added, and
further consideration will be given when RXIARG is being eliminated
(which is a part of the Ren/C plan to use pure Rebol headers for
extension DLLs vs. a mirrored type hierarchy).