TO-INTEGER/UNSIGNED and conversion fixes
This factors out the functionality of INTEGER!'s TO conversion into
a service function, and then adds a parameter to that function for
enforcing interpretation to be unsigned.  Then it creates a NATIVE!
version of TO-INTEGER! which provides an /UNSIGNED refinement
to use the new ability.

It also fixes a bug in the TO INTEGER! implementation for numbers
(introduced in fea5921).  The problem would occur when leading 0xFF
bytes were followed by a byte without the high bit set.  This would
lead to positive interpretations of those numbers.

Several other adjustments were made to correct edge cases or add
informative errors for surprising results.  As a by-product of writing the
routine for general purposes, it can generate a REBI64 directly into a
value cell or into a REBI64 without a value cell.