Bound Debug build DO step counter to MAX_U32
There's a counter used for debugging getting to a certain "tick count" of DO
steps in the source.  It will probe that step and let you set a breakpoint.  This
is very useful, however it didn't have a bound on it... so it would overflow.

Though "only a debug feature", running a debug build is supposed to be a
better-checked version of the release build.  So it shouldn't give any output
that wasn't requested.  This expands the range of where you set the tick
count by using an unsigned number and limits the tick counting range to
MAX_U32.