Print BINARY! as raw bytes to console, skip NONE!
An issue discovered regarding Rebol's limitations of interacting in a
web CGI environment is that it has no way to do full 8-bit transport to
the console.  Unlike WRITE, PRINT of a binary would not send the
raw bytes to the device but rather do an effective PROBE of the data.

The general idea that a PRINT of a value is a general-purpose way
of inspecting any arbitrary Rebol type is already undermined by its
longstanding behavior with BLOCK!.  Hence, in order to make it
possible to access things like terminal escape codes and other needs
it makes more sense to treat printing of a binary in a way that gives
full raw access to the device.

(This does not change the behavior of BINARY! inside of a block that
is passed to PRINT, which is currently the domain of FORM.)

This adds another behavior which has been sought after and requested,
which is that NONE! and UNSET! omit any print whatsoever.  This
makes it possible for a function call to "opt out" of a print.

Switch in `<r3-legacy>` for old behavior: PRINT-FORMS_EVERYTHING