Patch size conversions from binary streams to unsigned
Some size conversions of binary were affected by the sign
interpretation change...because it was being used in TLS and
HTTP to decode streams.  Hence the binaries were not full 64-bit
integers, and could be interpreted as signed when they were meant
to be unsigned.

This was obviously going to be a problem for clients not using
`<r3-legacy>`, which was known.  But for reasons of oversight and
false security from not getting alerts raised by conversions in the
tests, Mezzanine instances were not thought to be examined
manually.