Tighten/clarify UTF8 char decoding interfaces
This is a rescue of some tinkering from before Ren/C in a mostly
irrelevant clone.  It seemed to be clarifying enough to be worth going
ahead and trying to merge in before throwing that repository away.

The interface of Decode_UTF8_Char and its callers was complex
and obscured the need to handle certain invalid return results.  Some
places did not check the result and shouldn't have to.  This makes it
more standardized by matching the Scan_XXX routines more
closely.  It also adds documentation for the behavior.

A special error is also added to help distinguish invalid UTF8 data
from one with a codepoint simply too high for the Rebol implementation
to currently handle.

A likely redundant check is moved for now to assert that the redundant
check matches the result if the check is not run.