Use unsigned accumulator in enbase routines
This was another undefined behavior from left-shift caught by UBSan:

http://stackoverflow.com/questions/3784996/

It might appear that the accumulator being signed or unsigned would
not matter as it's just using the low 3 bytes.  But with undefined behavior
it is not required to get any of the bits correct if you shift any of the data
out of the range.  It *could* crash or fail arbitrarily beyond just getting the
number wrong.

There was no reason for the accumulator to be signed so this just makes
it unsigned in the enbase routines.