Add "SAVE/UNSAVE_VALUE" (rename to GUARD...)
In order to protect a series from being garbage collected in a way that was
efficient and safe for recursively protecting and unprotecting the same
series, the SAVE_SERIES and UNSAVE_SERIES macros existed.  They
used a stack to efficiently save and release the pointers, as well as to
check in the debug build to make sure the last saved series was the next
to be unsaved.

No parallel way of saving a value existed, perhaps because most values
live in series.  (You don't want to hold onto pointers to them because the
series can have their data arrays moved during reallocation.)  However, if
a value is known to live in stable memory it could use the same technique
to keep any series it references protected from GC (as long as it is, actually
stable).

Because no mechanism existed, it was necessary to put even a single value
into a locked series and then save *that* series.  This implements a light
solution just like SAVE_SERIES for SAVE_VALUE.  Additionally, it provides
a debug check to make sure that the saved value does not live inside of
series memory (which is too costly to run in general cases, but good to do
for peace of mind...and applicable to other routines that do not want to be
holding onto potentially unstable memory cells).

This also renames the macros from SAVE_XXX / UNSAVE_XXX to
PUSH_GUARD_XXX / DROP_GUARD_XXX.  This helps emphasize the
stacklike nature at the callsite, and "guarding" is a less ambiguous word for
what is being done to the series than "saving" (which might suggest some
kind of serialization).