Defer c-do.c assert on ROUTINE!, manage struct
All series created must be either freed or delegated to the memory
manager before the end of a function.  A series seemingly intended
to represent a struct element to be handled by the GC wasn't being
freed, so added the presumed missing MANAGE_SERIES().

Also commented out an aggressive assert, because the condition it
checks is going to be changed soon...so better to comment the assert
out than modify the code that didn't set that.