Kill parse longjmp hack, native THROW/NAMEs
Rebol uses C's precarious longjmp construct to implement error
handling, because it wants to (for instance) not have to add a check
after every memory allocation to determine if it failed...and then have
to gracefully find a way to move an error into its returning slot.

In PARSE's C implementation, it was using the error longjmp in a
hacky way in order to proxy out return results or throws from within an
embedded rule.  No error had actually occurred... it just was somewhat
deep into a stack and did not have a return result to indicate a throw
had happened.  Rather than adding the result, it pretended an error
occurred and "jumped the stack".

There are several reasons not to do this, including the unnecessary set
up and tear down of a setjmp/longjmp state on each parse call.  It also
creates a surprise on an error boundary that thinks an error means an
in-progress throw may be uncaught.  Yet erasing the thrown value for
parse on error would lose the very trick it was trying to do with raising
the error in the first place.

This resolves it by just removing the hack, and using a THROWN_FLAG
index (in the same way that the Do evaluator does) to bubble out from a
deep parse stack.  For its own RETURN keyword, it effectively throws to
itself using a throw identified w/name of the PARSE native function value.

Then this reinstates the assertions which were originally designed for
managing the side-value on the stack to carry the thrown argument.  It
also switches the names of the throw names for BREAK/CONTINUE/QUIT
and others to be the names of their natives.  This reopens the use of any
word for the user to throw and catch without worrying about stepping on
something the system is doing.