Add CONTINUE/WITH, modify BREAK/WITH, cleanup
This commit adds a /WITH refinement to CONTINUE, such that it
will skip the rest of the loop body but still act like it had returned the
value it is passed:

    >> map-each x [1 2 3] [continue/with 4 "foo"]
    == [4 4 4]

This also reconciles the behavior of /WITH for BREAK in a similar
way, so that it will effectively make the body return that value for the
last iteration.  This leads to a concrete difference from /RETURN
which did not have the intention (and would not make sense to put
onto continue).

For more explanation and why /RETURN is not being kept around
(instead suggesting people use CATCH+THROW), see porting guide:

https://trello.com/c/cOgdiOAD

An `<r3-legacy>` switch is added for making /WITH act with the old
behavior of /RETURN, which is used as a synonym...however the
two behaviors cannot be left running concurrently.  Roughly speaking
this means that if there are any BREAK/WITHs inside of MAP-EACH
in the mezzanine then it would break legacy mode.  There aren't
any (yet) so that's not being treated as priority.

This also includes some reorganization and cleaning of the loop
constructs.  That was part of a global audit of THROWN() boundaries
that triggered the observation behind this change.