Codebase-wide THROWN() audit, plus checks
Rebol has two methods of interrupting control: errors and throws.
The error mechanism makes use of setjmp/longjmp and must be
explicitly caught by certain stack levels, but a thrown value can be
passed back as a return value from any call into the evaluator and
needs to have some kind of handling at each evaluator call site...
either to catch it, to bubble it up through the next return result, or
to raise an error on it.

Despite the importance of each call into the evaluator having
explicit handling for throws, there was a tendency in the original code
to often forget to handle it.  It has been an ongoing process to give
routines names (like XXX_Throws()) to signal that the caller must
be responsible for triaging a throw.

The elimination of the "parse longjmp hack" facilitated the addition of
more checks, and a further realization that the "thrown bit" could even
be moved off of values and fully as a property of the stack.  This goes
through each instance of throwing and vets it.  While many bugs where
potential throws/returns/quits have been found over time, this catches
many more.

A potentially temporary tool of having natives and actions return a flag
when they are throwing a value or passing it on would be able to
replace the idea of carrying thrown bits on values themselves.  This is
generally desirable, because putting a thrown bit on the value makes it
easy to get lost if not checked.  But other mechanisms may arise that
are less manual and still not costly, so it may be that this is not kept in
the long term.