Fix bug introduced to AJOIN on nested molds
Commit 21ee09b introduced a bug, when it seemed that there was
an unnecessary use of the data stack.  The data stack was not needed
for protection of the values from GC it turns out...but rather to avoid a
potential recursive use of the MOLD buffer.

This resolves the issue, comments it, and modifies the code to be slightly
more obvious.