Global audit of DO of arrays vs. series at positions
This fixes a category of bugs that were mostly introduced into Ren/C
early on, by an apparent oversight on the distinction between Do_Blk()
and DO_BLK() while reorganizing Do.  They went unnoticed because
it is relatively infrequent to use blocks that aren't at their head position
for various purposes of execution (such as the body of an IF).

All call sites were reviewed, leading to simplification and correctness
on the issue.  Nomenclature and comments are also adjusted a bit more
to help readers understand the code better.