Add RELAX and OPTIONAL natives, OPT shorthand
This creates two natives that had been suggested for a while but not
named.  RELAX turns unsets into NONE! and leaves all otehr value
types alone, while OPTIONAL does the opposite.

OPT has been used variously as a variable name in places that are not
dialects like PARSE.  However, this adds it as an abbreviation for the
OPTIONAL in order to allow that to be experimented with.