Default IF, WHILE, EVERY to unsets, x: () legal
**WARNING** this change can cause subtle bugs in previous
patterns of the form:

    all [1 (if 1 > 2 [3]) 4]

While ALL always considered unsets a method for "opting out" of
the vote and not causing a test to fail, failing IF's NONE *would*
cause a test to fail.  New patterns of thinking a condition failing means
a 'void' or absence of data instead of a chainable "falsity" is a cool
interpretation--and worth the change systemically.  However, such
code will need to be ported to something more like:

    all [1 all [(1 > 2) 3] 4]

Or even more elegantly flattened to:

    all [1 (1 > 2) 3 4]

The option to switch to EITHER with NONE also exists, or to
experiment with the new RELAX native.

`<r3-legacy>` has the added switches:

    none-instead-of-unsets: false
    cant-unset-set-words: false