error trap in decompress, no manuals check in DROP_TRAP
The modification to use stock Zlib and its stream-based decompression
meant performing Rebol allocations to do generalized DEFLATE decode
with no pre-known size.  inflateInit() allocated internal state which needed
to be freed in inflateEnd() but this end would not be called if a longjmp
happened due to raising an error.

This spurred the loosening of the requirement of all manuals to be free
at the time of a DROP_TRAP; because the trapping had little to do with
the contract between the function and the caller...which wished in this case
to return a manually managed series.  The constraint seems to ultimately
be too restrictive considering the total number of checks for manual leaks
already in place in function dispatching.