FUNC + CLOS => native, non-head spec/body, cleanups
This makes FUNC and CLOS natives, and does a thorough refactor
of the creation mechanics.  The overall improvements lead to a speed
increase of about 60% in optimized builds in func or clos creation.

(Note: The motivation was not to increase the speed, rather to mitigate
what would be an impending slowdown in speed for definitional return if
the two were not made native.)

It corrects a bug by which regardless of the series position passed in
for the spec or body components in a MAKE FUNCTION! call, the series
would be used at their head.  This involves taking out the "optimization"
of not copying the series used in MAKE FUNCTION! .  Because FUNC
is targeting the addition of a definitional RETURN that will *not* be
provided by MAKE FUNCTION! (which will only have non-definitional
EXIT), calling FUNC is the go-to way of ensuring that the generated call
will have RETURN available vs. writing it oneself in the generator.

Because the copying nature of the primitive means user code cannot
get access to modify a body or spec while running, an additional
optimization is made that empty spec or body series are not copied but
rather usages of a single global empty series.  (Hence DOES need not
copy its spec block from `func [] body`.

The factoring to make the natives "clean" involved splitting out the
REB_COMMAND code better.  Additional documentation notes are in
the comments about definitional returns and the nature of COMMAND!.
The changes make the code *significantly* clearer.