Use Rebol REB_XXX type enum for Make Dispatch
A REBCNT was being used where specifically a type like REB_BLOCK
or REB_ERROR was expected.  This clears that up, as well as resolving
a potential bug in methodized construction (if it were to begin writing
the output value, do an evaluation, and then consult the make spec
after that...it could have been GC'd).