Patch molding and leak in TRACE
TRACE was invoking mold on an unmanaged series for debug
purposes, and hadn't been used recently to notice that this was
triggering an assert.

This works around the issue, although the source of the problem is
that series are being placed into REBVALs when they don't need to
be (which is the strategy for guessing that they should be implicitly
be made managed by the GC, and hence shouldn't be done with any
series that doesn't *need* to be in a value (and hence could be used
with Free_Series to avoid taxing the GC).