Patch tmp-boot.h generator for no tabs, all if => case
The tmp-boot.h generator had an issue with long names because it was
trying to tab things to a fixed alignment for comments...a too long name
jammed up against the column and created a syntax error.

This removes that, fixes some any [if... constructs that needed to be case
statements, and adds a comment header format fix to Routine dispatch.