Unite function copying + cloning, ask probing questions
In R3-Alpha this input:

    foo: func [x] [state: [] append state 'a print [x mold state]]
    foo 1 foo 2
    bar: copy :foo
    bar 3 bar 4
    foo 5

Will produce:

    1 [a]
    2 [a a]
    3 [a a a]
    4 [a a a a]
    5 [a a a]

The idea is that a function accumulates state that can be snapshot
in the form of "whatever you would get by deep copying the body".
However if a CLOS is used instead of a FUNC, you get:

    1 [a]
    2 [a]
    3 [a]
    4 [a]
    5 [a]

This commit coheres two implementations of the same copy operation--
one was a "copying" interface that moved source to target, while another
was a "clonifying" interface (my made-up-word) which expects the identical
copy to already be made and to reify it into a full fledged clone.  It also notes
that a suspicion mentioned about closures was true, and now only performs
the behavior on function copies.

Regardless of the long term choice in the matter, this reduces the total code...
adds comments...and clarifies.