Remove assertion from path code for now
During the thrown audit, it crossed into the path dispatch code and had
to punt a bit on it, because path dispatch needs a much bigger review.
An assertion regarding the thrown status of pvs.value that had been
holding up in debug builds is appearing to trigger in release builds on
setting the time the first time through a run:

    t: now
    t/time: 10:20:03

Note left in the source for when a fuller path dispatch review comes up.