Omit pure locals from HELP strings
The RETURN: was showing up in HELP because it was willing to
put ANY-WORD! in the list of arguments.  This filters it to the legal
types that can appear in specs.

It also adds a stray comment pending about why FUNC might be
more useful sometimes than MAKE FUNCTION! even when
transparent...