match CLOSURE! in fake spec, comment body lie
The fake function body generator was saying it ran MAKE on a
FUNCTION! even in the closure case.  So this needed another slot to
substitute in for the type.

While doing the modification, it seemed informative to throw in a comment
that the body was optimized and running "as if".  That would help inform
anyone who comes along and is puzzled why when they write the same
code that it is slower.

Also used #TYPE and #BODY instead of $1 / $2 as the substitution slots.