Undo datatype and word comparison equivalence
I read about an idea that was floating around on AltME that because it
was hard to SWITCH on actual DATATYPE! values, that it might be good if
the datatype and the word for that datatype would compare equal.  Hence
if you wrote `type-of 3 = 'integer!` it could work.

It seemed like not such a bad idea at first, and since people were already
doing TO-WORD conversions on types to compare them it would just be
a "little lie", and not apply in strict-equal!.

However, a much better alternative came along: the idea of making switch
able to do evaluations if you use a GET-WORD!, GET-PATH!, or PAREN!.
This genuinely useful feature covered the same desire in a less hacky way.

So this backs out that change.