Back out internal GROUP! and PAREN! equivalence
Due to the change that made switch able to escape into evaluation,
the work to try and make GROUP! and PAREN! equal came out looking
kind of hackish after that.  So this removes the change.

The battle for which concept owns the word in a TO-WORD sense can
wait for another time, as source code can be written to mention GROUP!
using this method...as long as the `switch type-of x [:some-type! [...]]`
pattern is used.