Fix case-sensitivity in string sort, support unicode, #2170
(Lingering incomplete fix from September 13, 2014 that had died as a
PR to GitHub rebol/rebol...being cleaned out during prep for source code
format conversion.)

In Rebol2:

    >> sort "ABCabcdefDEF"
    == "AaBbCcdDeEfF"

    >> sort/case "ABCabcdefDEF"
    == "ABCDEFabcdef"

In R3-Alpha:

    >> sort "ABCabcdefDEF"
    == "ABCDEFabcdef"

This fixes the behavior to be consistent with Rebol2.  It also adds support
for unicode characters, which were being "randomly" compared based on one
of their bytes.

To reduce the number of comparison functions that would have to be written
and avoid the use of global state, this imported the qsort_r implementation of
the quicksort instead of qsort.  (Fun fact: qsort_r() is not part of any standard...
so the parameter orders vary.  Rebol builds in the same version used by
Android Bionic.)