REB_END value non-zero, write header of end only
Rebol byte-size string series are zero-terminated (latin1 strings) as are
Rebol wide-char strings terminated by two zero-bytes (UCS-2).

Similarly, the REBVAL-holding series for ANY-ARRAY! have a terminal
value.  As with string iterations, it is possible to iterate positions inside
a series without passing around a size or index.  The value 0 was chosen
so that it would work easily with a zero fill, which has some notable
disadvantages in terms of obscuring when terminator management is
not being done consciously.

This changes the value to 1, which uncovered a few places that would
likely yield bugs under slight code arrangement--due to implicit termination.
Zero is then retaken to use for a debug fill value (the REB_TRASH).

Since there is now no generic function that can terminate either an array
or a "sequence" (non-array) series, either a test is required for a generic
routine...or the caller explicitly saying which they wish to terminate.  This
commit errs a little on the side of "too many routines" to avoid the test for
now...because it's best to avoid the test if you can.

Having this in place as well as the assertions supporting conscious control
of the terminator opens up a possibility.  It is not necessary to fill all the
bytes of a REBVAL in order for it to serve as a terminator...only the header.
So if *just* the header is marked (not a full value filled) it could be possible
to store a 2 element series in a byte arrangement that was just one element
and one element-header long.  This has an interesting possibility of being
able to put such a series directly into a series node itself...saving the data
allocation for both 0 and 1 element series.