Add <local> tag to FUNC + CLOS, move <infix>
With "Pure Local" generation offered via set-words, it has the problem of
being a bit "undifferentiated-looking":

    foo: func [date [date!] out: d1: d2:][
        out: [0 0]
        set [d1 d2] get-iso-year out/2: date/year

vs.

    foo: func [date [date!] /local out d1 d2][
        out: [0 0]
        set [d1 d2] get-iso-year out/2: date/year

The advantage of "pure locals" is that they aren't shown by WORDS-OF
and cannot be injected (e.g. by calling foo/local).  But the refinement lets
you use regular words and looks a little bit "less like code".

This modifies the FUNC generator to use the <local> tag as a cue to
convert WORD! following it into SET-WORD!, up to the next refinement.  It
will ignore any pure locals (set-words), and error on other kinds of words
(e.g. `<local> 'x #y :z`).  The appearance is better, and it gives "keyword-like
behavior" from a construct implementable entirely in userspace.

    foo: func [date [date!] <local> out d1 d2][
        out: [0 0]
        set [d1 d2] get-iso-year out/2: date/year