Do not overwrite pure local arg slots in APPLY
The pure-local skipping code in Apply_Block skipped the *parameter*, but did
not skip the *slot*.  Hence it would wind up overwriting definitional returns with
a remainder fill.

This fixes both the missing skip, and the remainder fill.  This repairs the
read http:// logic (but not https:// which seems to have a different problem)