Evaluate parens in refinement dispatch, ~10% speedup
The previous semi-hack of an evaluation trick so that get-words would
evaluate in paths has been upgraded to a much less hacky trick.  This
now allows for the evaluation of parens to be used to get the words to
use as refinements in dispatch:

    >> append/(first [only])/(to-word "dup") [a b c] [d e] 2
    == [a b c [d e] [d e]]

The method by which this is done--as well as reducing the number of
writes done into call frames--has improved performance in a build of
the hostilefork.com using the static builder by 10%.  This means Ren/C
optimized non-debug builds are--for the first time--are ever so slightly
faster than the rebolsource download (in that measurement).